Move "leave room" logic to separate controller
Part of general refactoring
