+++ matching-service/index.js
@@ -4,6 +4,7 @@ import { createServer } from 'http';
 import { Server } from "socket.io";
 import { ormCreatePendingMatch, ormClaimFirstMatchByDifficulty, ormFlushPendingMatchById } from "./model/pendingMatch-orm.js"
 import { v4 as uuidv4 } from 'uuid';
+import { addLeaveRoomCallback } from "./controller/leaveRoomController.js";
 
 const app = express();
 app.use(express.urlencoded({ extended: true }))
@@ -47,37 +48,10 @@ io.on("connection", (socket) => {
             
             const otherSocket = io.sockets.sockets.get(avaliableMatch.socket_id); // TODO: Detect if socket has already disconnected
             const room_id = uuidv4();
-
-            const leaveRoomCallbackFactory = (socket) => () => {
-                for (const room of socket.rooms) {
-                    if (room !== socket.id) {
-                        socket.to(room).emit("user_leaves", username);
-                        socket.leave(room);
-                        //Future proof for rooms with more than 2 people
-                        const room_contents = io.sockets.adapter.rooms.get(room);
-                        const room_occupancy = room_contents.size;
-                        if (room_occupancy < 2) { 
-                            const lonely_socket_id = room_contents.values().next().value;
-                            const lonely_socket = io.sockets.sockets.get(lonely_socket_id);
-
-                            lonely_socket.removeAllListeners("disconnecting");
-                            lonely_socket.removeAllListeners("leave_room");
-
-                            lonely_socket.leave(room);
-                            lonely_socket.emit("room_info", "Room has been torn down")
-                        }
-                    }
-                }
-            };
-            const addLeaveRoomCallback = (socket) => {
-                const leaveRoomCallback = leaveRoomCallbackFactory(socket)
-                socket.on("disconnecting", (_) => {leaveRoomCallback();});
-                socket.once("leave_room", leaveRoomCallback);
-            }
             
             //Handle other socket
             otherSocket.join(room_id); 
-            addLeaveRoomCallback(otherSocket)
+            addLeaveRoomCallback(io, otherSocket, avaliableMatch.username)
             socket.to(room_id).emit("matchSuccess","Found"); //TODO: Require ack, change eventName to match_result
             otherSocket.on("message", 
                 (msg) => otherSocket.to(room_id).emit("message", msg)
@@ -86,7 +60,7 @@ io.on("connection", (socket) => {
 
             //Handle this socket
             socket.join(room_id);
-            addLeaveRoomCallback(socket)
+            addLeaveRoomCallback(io, socket, username)
             socket.emit("matchSuccess","Found"); //TODO: Require ack, change eventName to match_result
             socket.on("message", 
                 (msg) => socket.to(room_id).emit("message", msg)
